pyproject.toml
[project]
name = "sise-claw"
version = "0.1.0"
description = "AI assistant liant LLM et serveur MCP Google Workspace"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.0.14",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "mistralai>=1.0.0",
]
